Updated screen and loop
- Added button to refresh day schedule
- Added button to refresh week schedule
- Seperation of files
- Deleted unnecesary imports
